refactor: move create_agent_environment to agent module
