+++ src/dfx/src/commands/fix.rs
@@ -3,8 +3,8 @@ use tokio::runtime::Runtime;
 
 use crate::{
     lib::{
-        environment::Environment, error::DfxResult, migrate::migrate,
-        provider::create_agent_environment,
+        agent::create_agent_environment, environment::Environment, error::DfxResult,
+        migrate::migrate,
     },
     NetworkOpt,
 };
