+++ src/dfx/src/commands/generate.rs
@@ -1,9 +1,9 @@
+use crate::lib::agent::create_anonymous_agent_environment;
 use crate::lib::builders::BuildConfig;
 use crate::lib::environment::Environment;
 use crate::lib::error::DfxResult;
 use crate::lib::models::canister::CanisterPool;
 use crate::lib::models::canister_id_store::CanisterIdStore;
-use crate::lib::provider::create_anonymous_agent_environment;
 use crate::NetworkOpt;
 
 use clap::Parser;
